formatting

+++ src/main/java/org/object/Platform.java
@@ -8,16 +8,17 @@ public class Platform extends Sprite {
 
     public Platform(float posX, float posY, float width, float height) {
         super(posX, posY);
-
         this.width = width;
         this.height = height;
     }
 
     public void render(Graphics g) {
         g.setColor(new Color(120, 80, 50));
-        g.fillRect((int) (posX - width / 2) - (int) Renderer.camX + (int) Renderer.gameWidth / 2,
-                (int) (posY - height / 2) - (int) Renderer.camY + (int) Renderer.gameHeight / 2, (int) width, (int) height);
+        g.fillRect(
+                (int) (posX - width / 2) - (int) Renderer.camX + Renderer.gameWidth / 2,
+                (int) (posY - height / 2) - (int) Renderer.camY + Renderer.gameHeight / 2,
+                (int) width,
+                (int) height
+        );
     }
-
-
 }
